Add support for a targetPlayers list on Effect also took a big pass on cleaning up important issues noted by findbugs

+++ src/main/java/de/slikey/effectlib/effect/LineEffect.java
@@ -71,7 +71,7 @@ public class LineEffect extends Effect {
         } else {
             target = getTarget();
         }
-        double amount = particles / zigZags;
+        int amount = particles / zigZags;
         if (target == null) {
             cancel();
             return;
